requirements.txt
Django==3.0.5
django-heroku==0.3.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
pexels-api==1.0.1
pipenv==2018.11.26
pygal==2.4.0
spotipy==2.12.0
boto.s3.connection